requirements.txt
fastapi
uvicorn[standard]
pydantic
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
reportlab
python-dotenv